+++ robot/resources/python/my_py_lib_to_py_test/_setup.py
@@ -2,7 +2,7 @@ import sys
 import os
 from pathlib import Path
 
-project_root = Path(__file__).parent.parent.parent
+project_root = Path(__file__).parent.parent.parent.parent
 sys.path.insert(0, str(project_root))
 
 project_cache = Path(__file__).parent.parent.parent / '.webdriver_cache'
